requirements.txt
# Core dependencies
langchain-google-genai==2.0.4
pydantic==2.11.0
pymongo>=4.9,<4.10
motor==3.6.0

# Web framework
fastapi==0.115.4
uvicorn==0.32.0
websockets>=13.0.0

# Utilities
python-dotenv==1.0.1
